refactor: reorganise source files
