+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/datasource/SpanDataSource.kt
@@ -1,5 +1,6 @@
-package io.embrace.android.embracesdk.arch
+package io.embrace.android.embracesdk.arch.datasource
 
+import io.embrace.android.embracesdk.arch.destination.SpanEventData
 import io.embrace.android.embracesdk.internal.spans.SpanService
 import io.embrace.android.embracesdk.spans.EmbraceSpan
