+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/datasource/SpanDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.arch
+package io.embrace.android.embracesdk.arch.datasource
 
 import io.embrace.android.embracesdk.arch.limits.LimitStrategy
 import io.embrace.android.embracesdk.internal.spans.SpanService
