+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/destination/LogWriter.kt
@@ -1,0 +1,6 @@
+package io.embrace.android.embracesdk.arch.destination
+
+/**
+ * Placeholder type for the log service.
+ */
+internal interface LogWriter
