+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/destination/SpanAttributeData.kt
@@ -1,0 +1,9 @@
+package io.embrace.android.embracesdk.arch.destination
+
+/**
+ * Represents a span attribute that can be added to the current session span.
+ */
+internal class SpanAttributeData(
+    val key: String,
+    val value: String
+)
