+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/destination/SpanEventData.kt
@@ -1,0 +1,10 @@
+package io.embrace.android.embracesdk.arch.destination
+
+/**
+ * Represents a span event that can be added to the current session span.
+ */
+internal class SpanEventData(
+    val spanName: String,
+    val spanStartTimeNanos: Long,
+    val attributes: Map<String, String>?
+)
